requirements.txt
gdown
tensorflow==2.18.0
keras==3.8.0
opencv-python==4.11.0
scikit-learn==1.6.1
matplotlib==3.10.0
pillow